requirements.txt
scitools-iris==3.7.0
cf-python==3.16.0
netcdf4==1.6.5
xarray==2024.2.0
nctoolkit==1.1.6